requirements.txt
absl-py==2.0.0
aiohttp==3.9.0
aiosignal==1.3.1
asttokens==2.4.1
attrs==23.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
flatbuffers==23.5.26
fonttools==4.44.3
frozenlist==1.4.0
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
kiwisolver==1.4.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
mediapipe==0.10.8
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==23.2
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
platformdirs==4.0.0
prompt-toolkit==3.0.41
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
sounddevice==0.4.6
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
urllib3==2.1.0
wcwidth==0.2.10
yarl==1.9.2
